src/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.0
python-dotenv==1.0.0
openai==1.3.0
tweepy==4.14.0
praw==7.7.1
requests==2.31.0
python-dateutil==2.8.2
aiohttp==3.9.1